backend/requirements.txt
openai-whisper
ffmpeg-python 
#npm install react-router-dom
fastapi
uvicorn
pytubefix